Stop caching instance methods.
Checking files on every exception is not that expensive and it stops us
from having to bounce the server whenever we add/remove a template.

I'm not completely sure why the bouncing was necessary. Seems that the exception
would be fresh each time. But perhaps I am missing something so just removing
caching for now.